% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfrecords.R
\name{get_class_and_type}
\alias{get_class_and_type}
\title{Get classes and value types from an object supported.}
\usage{
get_class_and_type(x)
}
\description{
Get classes and value types from an object supported.
}
