% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfrecords.R
\name{write_tfrecords}
\alias{write_tfrecords}
\title{Write tfrecords from a list of data.}
\usage{
write_tfrecords(data, path)
}
\arguments{
\item{data}{list of data.}
}
\description{
Write tfrecords from a list of data.
}
\examples{
data <- list(
 x = matrix(1:1000, nrow = 100),
 y = matrix(1:1000/2.1, nrow = 100),
 z = Matrix::rsparsematrix(nrow = 100, ncol = 10, density = 0.3)
 )
 
 write_tfrecords(data, "example.tfrecords")

}
