% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfrecords.R
\name{write_tfrecords}
\alias{write_tfrecords}
\title{Write tfrecords from a list of data.}
\usage{
write_tfrecords(data, path)
}
\arguments{
\item{data}{list of data.}
}
\description{
Write tfrecords from a list of data.
}
\examples{
data <- list(
 x = matrix(1:1000, nrow = 100),
 y = matrix(1:1000/2.1, nrow = 100),
 z = Matrix::Matrix(sample(0:1, prob = c(0.9, 0.1), size = 1000, replace = TRUE), nrow = 100)
 )


}
